+++ examples/simple-template/src/customConfig/custom/fieldDataCollection/index.js
@@ -1,3 +1,4 @@
 export * from './fieldDataNotice';
+export * from './fieldDataUser';
 export * from './fieldDataWorkflowCase';
 export * from './fieldDataWorkflowCaseFormAttachment';