+++ examples/simple-template/src/modelBuilders/custom/flowCase.js
@@ -1,5 +1,6 @@
 import {
   getTacitlyState,
+  pretreatmentRemoteListData,
   pretreatmentRemotePageListData,
   pretreatmentRemoteSingleData,
   reducerCollection,
@@ -15,6 +16,7 @@ import {
   pageListWaitApproveData,
   passData,
   refuseData,
+  singleListNextNodeApproverData,
   submitApprovalData,
   submitFormData,
 } from '../../services/flowCase';
@@ -23,6 +25,7 @@ export const flowCaseTypeCollection = {
   pageList: 'flowCase/pageList',
   pageListWaitApprove: 'flowCase/pageListWaitApprove',
   pageListLatestApprove: 'flowCase/pageListLatestApprove',
+  singleListNextNodeApprover: 'flowCase/singleListNextNodeApprover',
   get: 'flowCase/get',
   createFlowCase: 'flowCase/createFlowCase',
   submitForm: 'flowCase/submitForm',
@@ -104,6 +107,32 @@ export function buildModel() {
         const response = yield call(pageListLatestApproveData, payload);
 
         const dataAdjust = pretreatmentRemotePageListData({
+          source: response,
+          successCallback: pretreatmentSuccessCallback || null,
+          failCallback: pretreatmentFailCallback || null,
+        });
+
+        yield put({
+          type: reducerNameCollection.reducerRemoteData,
+          payload: dataAdjust,
+          alias,
+          ...reducerDefaultParameters,
+        });
+
+        return dataAdjust;
+      },
+      *singleListNextNodeApprover(
+        {
+          payload,
+          alias,
+          pretreatmentSuccessCallback,
+          pretreatmentFailCallback,
+        },
+        { call, put },
+      ) {
+        const response = yield call(singleListNextNodeApproverData, payload);
+
+        const dataAdjust = pretreatmentRemoteListData({
           source: response,
           successCallback: pretreatmentSuccessCallback || null,
           failCallback: pretreatmentFailCallback || null,