+++ examples/simple-template/CHANGELOG.md
@@ -1,5 +1,18 @@
 # taro-fast-framework-template
 
+## 1.0.41
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - taro-fast-component@3.0.45
+  - taro-fast-framework@3.1.39
+  - taro-fast-common@3.1.30
+  - taro-fast-component-extra@3.0.45
+  - taro-fast-component-prism@3.0.46
+  - taro-fast-design-playground@3.0.32
+
 ## 1.0.40
 
 ### Patch Changes