+++ packages/taro-fast-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 3.1.30
+
+### Patch Changes
+
+- update
+
 ## 3.1.29
 
 ### Patch Changes